+++ src/dithering/NearestColor.cpp
@@ -4,7 +4,7 @@
 
 #include <cmath>
 
-int nearestColor(int bit, double pixel_color) {
+float nearestColor(int bit, double pixel_color) {
     if (pixel_color < 128)
         auto a = 1;
     int differenceBetweenColorLevels = std::pow(2, 9 - bit);
